+++ components/GraphicProject/GraphicProject.test.js
@@ -168,7 +168,7 @@ describe( '<GraphicProject />, for GraphQL data', () => {
     expect( modalImage.exists() ).toEqual( true );
     expect( modalImage.props() ).toEqual( {
       thumbnail: selectedUnit.url,
-      thumbnailMeta: { alt: selectedUnit.alt },
+      thumbnailMeta: { alt: normalizedData.alt },
     } );
   } );
 
@@ -209,12 +209,13 @@ describe( '<GraphicProject />, for GraphQL data', () => {
     expect( alt.exists() ).toEqual( true );
     expect( title.exists() ).toEqual( true );
     expect( title.text() ).toEqual( 'Alt (Alternative) Text:' );
-    expect( alt.contains( selectedUnit.alt ) ).toEqual( true );
+    expect( alt.contains( normalizedData.alt ) ).toEqual( true );
   } );
 
   it( 'renders the ModalPostMeta', () => {
     const postMeta = wrapper.find( 'modal-post-meta' );
     const { owner, projectType, published } = normalizedData;
+    const { language } = selectedUnit;
 
     expect( postMeta.exists() ).toEqual( true );
     expect( postMeta.props() ).toEqual( {
@@ -222,6 +223,7 @@ describe( '<GraphicProject />, for GraphQL data', () => {
       logo: 'image-stub',
       source: owner,
       datePublished: published,
+      textDirection: language.text_direction,
     } );
   } );
 
@@ -232,6 +234,52 @@ describe( '<GraphicProject />, for GraphQL data', () => {
     expect( postTags.props() ).toEqual( {
       tags: getTransformedLangTaxArray( props.item.categories ),
     } );
+  } );
+} );
+
+describe( '<GraphicProject />, for GraphQL data with no project alt', () => {
+  const props = {
+    displayAsModal: true,
+    isAdminPreview: true,
+    item: {
+      ...graphicGraphqlMock,
+      alt: '',
+    },
+    useGraphQl: true,
+  };
+
+  const normalizedData = normalizeGraphicProjectByAPI( {
+    file: props.item,
+    useGraphQl: props.useGraphQl,
+  } );
+
+  let Component;
+  let wrapper;
+  const selectedUnit = normalizedData.images.find( img => img.language.locale === 'en-us' );
+
+  beforeEach( () => {
+    Component = <GraphicProject { ...props } />;
+    wrapper = mount( Component );
+  } );
+
+  it( 'renders the ModalImage thumbnail', () => {
+    const modalImage = wrapper.find( 'modal-image' );
+
+    expect( modalImage.exists() ).toEqual( true );
+    expect( modalImage.props() ).toEqual( {
+      thumbnail: selectedUnit.url,
+      thumbnailMeta: { alt: normalizedData.title },
+    } );
+  } );
+
+  it( 'renders the alt text', () => {
+    const alt = wrapper.find( '.graphic-project__content.alt' );
+    const title = alt.find( '.graphic-project__content__title' );
+
+    expect( alt.exists() ).toEqual( true );
+    expect( title.exists() ).toEqual( true );
+    expect( title.text() ).toEqual( 'Alt (Alternative) Text:' );
+    expect( alt.contains( normalizedData.title ) ).toEqual( true );
   } );
 } );
 
@@ -437,6 +485,7 @@ describe( '<GraphicProject />, for ElasticSearch data', () => {
   it( 'renders the ModalPostMeta', () => {
     const postMeta = wrapper.find( 'modal-post-meta' );
     const { owner, projectType, published } = normalizedData;
+    const { language } = selectedUnit;
 
     expect( postMeta.exists() ).toEqual( true );
     expect( postMeta.props() ).toEqual( {
@@ -444,6 +493,7 @@ describe( '<GraphicProject />, for ElasticSearch data', () => {
       logo: 'image-stub',
       source: owner,
       datePublished: published,
+      textDirection: language.text_direction,
     } );
   } );
 
@@ -454,6 +504,51 @@ describe( '<GraphicProject />, for ElasticSearch data', () => {
     expect( postTags.props() ).toEqual( {
       tags: props.item.categories,
     } );
+  } );
+} );
+
+describe( '<GraphicProject />, for ElasticSearch data with no image alts', () => {
+  const props = {
+    item: {
+      ...graphicElasticMock[0]._source,
+      images: graphicElasticMock[0]._source.images.map( img => ( {
+        ...img,
+        alt: '',
+      } ) ),
+    },
+  };
+
+  const normalizedData = normalizeGraphicProjectByAPI( {
+    file: props.item,
+  } );
+
+  let Component;
+  let wrapper;
+  const selectedUnit = normalizedData.images.find( img => img.language.locale === 'en-us' );
+
+  beforeEach( () => {
+    Component = <GraphicProject { ...props } />;
+    wrapper = mount( Component );
+  } );
+
+  it( 'renders the ModalImage thumbnail', () => {
+    const modalImage = wrapper.find( 'modal-image' );
+
+    expect( modalImage.exists() ).toEqual( true );
+    expect( modalImage.props() ).toEqual( {
+      thumbnail: selectedUnit.url,
+      thumbnailMeta: { alt: normalizedData.title },
+    } );
+  } );
+
+  it( 'renders the alt text', () => {
+    const alt = wrapper.find( '.graphic-project__content.alt' );
+    const title = alt.find( '.graphic-project__content__title' );
+
+    expect( alt.exists() ).toEqual( true );
+    expect( title.exists() ).toEqual( true );
+    expect( title.text() ).toEqual( 'Alt (Alternative) Text:' );
+    expect( alt.contains( normalizedData.title ) ).toEqual( true );
   } );
 } );